requirements.txt
absl-py==2.1.0
altair==5.5.0
antlr4-python3-runtime==4.9.3
apturl==0.5.2
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.46
botocore==1.38.46
Brlapi==0.8.3
cached_path==1.7.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.4
colorlog==6.9.0
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
curl_cffi==0.11.4
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.18.0
Flask==3.1.1
flatbuffers==24.3.25
fonttools==4.58.4
fredapi==0.5.2
frozendict==2.4.6
fsspec==2025.5.1
gast==0.5.4
gdown==4.7.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.64.0
gyp==0.1
h5py==3.11.0
hf-xet==1.1.5
httplib2==0.20.2
huggingface-hub==0.33.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
hyfi==0.2.20
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.3.3
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==8.10.0
multitasking==0.0.11
namex==0.0.8
narwhals==1.44.0
netifaces==0.11.0
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
opt-einsum==3.3.0
optree==0.11.0
pandas==1.5.3
peewee==3.18.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.8
plotly==6.2.0
proto-plus==1.26.1
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==1.10.22
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
pytz==2025.2
pyxdg==0.27
PyYAML==5.4.1
referencing==0.36.2
reportlab==3.6.8
requests==2.32.4
rich==13.7.1
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
scipy==1.15.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.46.1
systemd-python==234
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.11.0
typing-extensions==4.11.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.35.0
wadllib==1.3.6
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.16.0
xdg==5
xkit==0.0.0
yfinance==0.2.64
zipp==1.0.0
fastapi==0.115.14